requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
decorator==5.1.1
dnspython==2.2.1
executing==0.8.3
Flask==2.1.2
idna==3.3
importlib-metadata==4.12.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
numpy==1.23.0
opencv-python==4.6.0.66
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pymongo==4.1.1
requests==2.28.1
six==1.16.0
stack-data==0.3.0
torch==1.12.0
torchvision==0.13.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
